pyproject.toml
[project]
name = "log_guard"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dask>=2023.5.0",
    "fastparquet>=2025.12.0",
    "hdbscan>=0.8.41",
    "kafka-python>=2.3.0",
    "nbformat>=5.10.4",
    "pandas>=2.0.3",
    "plotly>=6.5.0",
    "prometheus-client>=0.23.1",
    "psycopg2-binary>=2.9.11",
    "pyarrow>=22.0.0",
    "scikit-learn>=1.8.0",
]
